Split declaration resolver behavior tasks
