+++ src/typechecker/tests/declaration_validation/resolver_replay.rs
@@ -1,5 +1,6 @@
 use super::*;
 
+mod behavior_tasks;
 mod semantic_bundles;
 
 #[test]
@@ -51,91 +52,6 @@ Point: { x: i32 = 1 }
     assert!(matches!(
         type_reference_tasks.as_slice(),
         [ResolverTypeReferenceValidationTask::Struct { name: "Point", .. }]
-    ));
-}
-
-#[test]
-fn resolver_behavior_declaration_metadata_tasks_collect_only_behavior_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-main = () i32 { 1 }
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_behavior_declaration_metadata_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].name, "Json");
-}
-
-#[test]
-fn resolver_behavior_replay_task_helper_pushes_metadata_and_type_refs_together() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-"#,
-    );
-    let mut behavior_tasks = Vec::new();
-    let mut type_reference_tasks = Vec::new();
-
-    let handled = TypeChecker::push_resolver_behavior_replay_tasks(
-        &program.declarations[0],
-        &mut behavior_tasks,
-        &mut type_reference_tasks,
-    );
-
-    assert!(handled);
-    assert_eq!(behavior_tasks.len(), 1);
-    assert_eq!(behavior_tasks[0].name, "Json");
-    assert!(matches!(
-        type_reference_tasks.as_slice(),
-        [ResolverTypeReferenceValidationTask::Behavior { name: "Json", .. }]
-    ));
-}
-
-#[test]
-fn resolver_behavior_impl_replay_task_helper_pushes_metadata_and_type_refs_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) StaticString
-}
-
-Point.implements(Json) {
-    encode = (self: Point) StaticString { "point" }
-}
-"#,
-    );
-    let mut behavior_impl_tasks = Vec::new();
-    let mut type_reference_tasks = Vec::new();
-
-    let handled = TypeChecker::push_resolver_behavior_impl_replay_tasks(
-        &program.declarations[2],
-        &mut behavior_impl_tasks,
-        &mut type_reference_tasks,
-    );
-
-    assert!(handled);
-    assert_eq!(behavior_impl_tasks.len(), 1);
-    assert_eq!(behavior_impl_tasks[0].ast_type_name, "Point");
-    assert_eq!(behavior_impl_tasks[0].behavior, "Json");
-    assert!(matches!(
-        type_reference_tasks.as_slice(),
-        [ResolverTypeReferenceValidationTask::ImplBlock {
-            type_name: "Point",
-            ..
-        }]
     ));
 }
 
@@ -207,33 +123,4 @@ make = () Point { Point { x: 1 } }
         type_reference_tasks.as_slice(),
         [ResolverTypeReferenceValidationTask::Function { name: "make", .. }]
     ));
-}
-
-#[test]
-fn resolver_behavior_impl_block_declaration_tasks_collect_only_behavior_impls() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) StaticString
-}
-
-Point.impl = {
-    x_value = (value: Point) i32 { value.x }
-}
-
-Point.implements(Json) {
-    encode = (value: Point) StaticString { "point" }
-}
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_behavior_impl_block_declaration_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].ast_type_name, "Point");
-    assert_eq!(tasks[0].behavior, "Json");
-    assert_eq!(tasks[0].methods.len(), 1);
 }
